# docs/swagger.yaml
basePath: /rederinghub.io/v1
definitions:
...
  response.JsonResponse:
    properties:
      data: {}
      error:
        $ref: '#/definitions/response.RespondErr'
      status:
        type: boolean
    type: object
...
  response.RespondErr:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
...
info:
  contact: {}
  description: This is a sample server Generative.
  title: Generative.xyz APIs
  version: 1.0.0
paths:
...
  /project/{contractAddress}/{projectID}/token:
    get:
      consumes:
      - application/json
      description: Search Token
      parameters:
      - description: page
        in: query
        name: page
        type: string
      - description: limit
        in: query
        name: limit
        type: integer
      - description: search
        in: query
        name: search
        type: string
      - description: projectID request
        in: path
        name: projectID
        required: true
        type: string
      - description: contractAddress request
        in: path
        name: contractAddress
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.JsonResponse'
      summary: Search Token
      tags:
      - Search
...
  /search:
    get:
      consumes:
      - application/json
      description: Search
      parameters:
      - description: page
        in: query
        name: page
        type: string
      - description: limit
        in: query
        name: limit
        type: integer
      - description: search
        in: query
        name: search
        type: string
      - description: object type
        in: query
        name: type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.JsonResponse'
      summary: Search
      tags:
      - Search
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"